backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Document Processing
pypdf>=4.0.1
python-pptx>=0.6.23
python-docx>=1.1.0

# Vector Store & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.1

# LLM Integration
langchain>=0.1.4
langchain-community>=0.0.16

# HTTP Client
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
